Separated DetectingFace into two components for easier reuse
